requirements.txt
python-dotenv==1.0.0
google-generativeai==0.5.0
anthropic==0.25.1
PyAutoGUI==0.9.53